Refactor AddPromptsBookmarkModal
